+++ ChangeLog.md
@@ -9,6 +9,8 @@
     - Back on 'command' deprecation.
       Rather say that 'command' and 'commands' are both freely intercheangeable.
       May consider that 'command' would be used when a single command is expected, and 'commands' when several commands are possible.
+    - Remove 'stdinFromNull' option from TTP::commandExec() function - Introduced in v4.6, disabled in v4.13
+	  Rationale: https://metacpan.org/pod/Capture::Tiny#LIMITATIONS prevents against acting on standard filehandles before capture() call
 
 ### 4.16.1